+++ cmd/clone.go
@@ -28,16 +28,6 @@ var cloneCmd = &cobra.Command{
 
 func cloneFunc(cmd *cobra.Command, argz []string) {
 
-	if cmd.Flags().Changed("color") {
-		colorToggle := cmd.Flag("color").Value.String()
-		if colorToggle == "on" {
-			os.Setenv("GHORG_COLOR", colorToggle)
-		} else {
-			os.Setenv("GHORG_COLOR", "off")
-		}
-
-	}
-
 	if cmd.Flags().Changed("path") {
 		absolutePath := configs.EnsureTrailingSlash((cmd.Flag("path").Value.String()))
 		os.Setenv("GHORG_ABSOLUTE_PATH_TO_CLONE_TO", absolutePath)